refactor the ast file
